Temporary fix for bug in unspent tx selection
Temp fix for keeping the unspent balance and tx's in sync. The resulting bug before the commit was that more unspent tx's were included as inputs which would result in the intended tnx having a change/remainder value. Turns out the system/network allocates any tx change back to miners. Possibly only if you don't explicitly specify which wallet the change should go to. So this needs to be investigated further. I'm not working on this anymore but pushing my last version in case someone wants to build on this. If you do, please take a stab at actually designing this better. This was scripted in this manner only so I could figure my way through the RPC interaction and prove that managing nodes in this manner is in fact possible.